#24 - moved Issuers integration tests to IssuersIntegrationTest
